[SSHD-1038] Refactor packages from a module into a cleaner hierarchy
renamed:    sshd-ldap: org.apache.sshd.server.auth -> org.apache.sshd.ldap
renamed:    sshd-ldap: org.apache.sshd.common.util.net -> org.apache.sshd.ldap
renamed:    sshd-ldap: org.apache.sshd.server.auth -> org.apache.sshd.ldap
renamed:    sshd-ldap: org.apache.sshd.server.auth.password -> org.apache.sshd.ldap
renamed:    sshd-ldap: org.apache.sshd.server.auth.pubkey -> org.apache.sshd.ldap
